Move functions for dealing with ST to their own module
